+++ gems/quilt_rails/lib/quilt_rails/engine.rb
@@ -6,6 +6,10 @@ module Quilt
 
     config.quilt = Quilt.configuration
 
+    initializer(:initialize_quilt_logger, after: :initialize_logger) do
+      config.quilt.logger = ::Rails.logger
+    end
+
     initializer(:mount_quilt, before: :add_builtin_route) do |app|
       if config.quilt.mount?
         app.routes.append do