+++ gems/quilt_rails/test/test_helper.rb
@@ -7,8 +7,12 @@ require 'minitest/autorun'
 require 'rails'
 require 'mocha/minitest'
 require 'quilt_rails'
+require 'active_support/testing/deprecation'
 require_relative "./support/generator_test_helpers"
+
+Quilt.configuration.logger = Logger.new(nil)
 
 class ActiveSupport::TestCase
   include GeneratorTestHelpers
+  include ActiveSupport::Testing::Deprecation
 end